scripts/terraform/cloudfunctions/ce/src/python/requirements.txt
# Function dependencies, for example:
# package>=version
google.cloud
google.cloud.bigquery
google.cloud.storage